Post: How we do Android API calls
This is a more or less complete version of a post about a clever way I
came up with to solve some problems we were experiencing with API calls
on Android. I've since seen more than [one question] about how to do what
we did so I figured our first Android blog post was in order!

[one question]: https://plus.google.com/116631804903803974311/posts/bKEo8K7Gezd